refactor : Remove LottoNumbers object because duplicate with Lotto object
