+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -16,12 +16,10 @@ class LottoGameTest {
     public static final int TICKET_NUMBER = 5;
     private LottoGame lottoGame = new LottoGame();
     private List<LottoNumber> winnerNumbers = new ArrayList<>();
-    private LottoNumbers winnerLottoNumbers;
 
     @BeforeEach
     void setUp() {
         IntStream.rangeClosed(1, 6).forEach(number -> winnerNumbers.add(new LottoNumber(number)));
-        winnerLottoNumbers = new LottoNumbers(winnerNumbers);
     }
 
     @Test
@@ -39,7 +37,7 @@ class LottoGameTest {
             assertEquals(lotto.getAward(), Award.UNIDENTIFIED);
         }
 
-        lottoGame.drawLotto(winnerLottoNumbers);
+        lottoGame.drawLotto(winnerNumbers);
         for (Lotto lotto : lottoGame.getLottos().getLottos()) {
             assertNotEquals(lotto.getAward(), Award.UNIDENTIFIED);
         }
@@ -50,7 +48,7 @@ class LottoGameTest {
     @DisplayName("countWinners 함수를 호출했을 때 각각의 합이 ticket의 숫자와 같은 지 확인한다.")
     void countWinners() {
         lottoGame.purchase(TICKET_NUMBER);
-        lottoGame.drawLotto(winnerLottoNumbers);
+        lottoGame.drawLotto(winnerNumbers);
         long countAll = lottoGame.countWinners(Award.FIRST) +
                 lottoGame.countWinners(Award.SECOND) +
                 lottoGame.countWinners(Award.THIRD) +
