+++ src/test/java/lotto/domain/LottoTest.java
@@ -15,6 +15,7 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -22,7 +23,6 @@ class LottoTest {
 
     private List<LottoNumber> winnerNumbers = new ArrayList<>();
     private List<LottoNumber> lottoNumberList = new ArrayList<>();
-    private LottoNumbers lottoNumbers;
     private Lotto randomLotto;
     private Lotto sortedLotto;
 
@@ -30,23 +30,22 @@ class LottoTest {
     void setUp() {
         IntStream.rangeClosed(1, 6).forEach(number -> winnerNumbers.add(new LottoNumber(number)));
         IntStream.rangeClosed(1, 6).forEach(number -> lottoNumberList.add(new LottoNumber(number)));
-        lottoNumbers = new LottoNumbers(lottoNumberList);
         randomLotto = new Lotto(LottoListGenerator.createLottoNumbers());
-        sortedLotto = new Lotto(lottoNumbers);
+        sortedLotto = new Lotto(lottoNumberList);
     }
 
     @Test
     @DisplayName("생성한 로또의 숫자의 개수가 6임을 확인한다")
     void size() {
-        assertThat(randomLotto.getNumbers().size()).isEqualTo(LottoNumbers.SIZE);
-        assertThat(sortedLotto.getNumbers().size()).isEqualTo(LottoNumbers.SIZE);
+        assertThat(randomLotto.getNumbers().size()).isEqualTo(Lotto.SIZE);
+        assertThat(sortedLotto.getNumbers().size()).isEqualTo(Lotto.SIZE);
     }
 
     @Test
     @DisplayName("랜덤 로또를 생성했을 때 각 숫자들은 최소숫자 1 이상, 최대숫자 45 이하임을 확인한다.")
     void lottoWithMixStrategy() {
-        LottoNumbers lottoNumbers = randomLotto.getNumbers();
-        for (LottoNumber lottoNumber : lottoNumbers.getLottoNumbers()) {
+        List<LottoNumber> lottoNumbers = randomLotto.getNumbers();
+        for (LottoNumber lottoNumber : lottoNumbers) {
             assertThat(lottoNumber.getLottoNumber()).isGreaterThanOrEqualTo(LottoNumber.MIN_NUMBER);
             assertThat(lottoNumber.getLottoNumber()).isLessThanOrEqualTo(LottoNumber.MAX_NUMBER);
         }
@@ -55,10 +54,10 @@ class LottoTest {
     @Test
     @DisplayName("정렬 로또를 생성했을 때 각 숫자들은 1부터 6임을 확인한다.")
     void lottoWithSortStrategy() {
-        LottoNumbers lottoNumbers = sortedLotto.getNumbers();
+        List<LottoNumber> lottoNumbers = sortedLotto.getNumbers();
         for (int i = 0; i < lottoNumbers.size(); i++) {
             LottoNumber number = new LottoNumber(i + 1);
-            assertEquals(lottoNumbers.getLottoNumbers().get(i), number);
+            assertEquals(lottoNumbers.get(i), number);
         }
     }
 
@@ -71,9 +70,8 @@ class LottoTest {
         assertEquals(sortedLotto.getAward(), Award.UNIDENTIFIED);
 
         List<LottoNumber> winnerNumbers = parseStringNumbersToList(winnerNumberString);
-        LottoNumbers winnerLottoNumbers = new LottoNumbers(winnerNumbers);
 
-        sortedLotto.drawLotto(winnerLottoNumbers);
+        sortedLotto.drawLotto(winnerNumbers);
         assertEquals(sortedLotto.getAward(), Award.valueOf(type));
     }
 
@@ -92,7 +90,25 @@ class LottoTest {
     @CsvSource(value = {"FIRST:true","SECOND:false"}, delimiter = ':')
     @DisplayName("1등으로 당첨 로또의 isWinner 함수의 parameter로 First, Second를 넣었을 때 각각 ture, false를 리턴한다.")
     void isWinnerFirstAward(String type, boolean result){
-        sortedLotto.drawLotto(new LottoNumbers(winnerNumbers));
+        sortedLotto.drawLotto(winnerNumbers);
         assertEquals(sortedLotto.isWinner(Award.valueOf(type)), result);
+    }
+
+    @Test
+    @DisplayName("LottoNumbers 생성 시 리스트의 사이즈가 6이 아니면 IllegalArgumentException이 발생한다")
+    void sizeException() {
+        List<LottoNumber> smallSizeNumberList = new ArrayList<>();
+        IntStream.rangeClosed(1, 5).forEach(number -> smallSizeNumberList.add(new LottoNumber(number)));
+        assertThatIllegalArgumentException().isThrownBy(() -> new Lotto(smallSizeNumberList))
+                .withMessage("LottoNumbers의 사이즈가 잘못 입력되었습니다. 입력 사이즈 : " + smallSizeNumberList.size());
+    }
+
+    @Test
+    @DisplayName("중복된 숫자로 LottoNumbers생성 시 IllegalArgumentException이 발생한다.")
+    void duplicateNumbers(){
+        List<LottoNumber> duplicateNumbers = new ArrayList<>();
+        IntStream.rangeClosed(1, 6).forEach(number -> duplicateNumbers.add(new LottoNumber(1)));
+        assertThatIllegalArgumentException().isThrownBy(() -> new Lotto(duplicateNumbers));
+
     }
 }
